restructure code to fix import cycles in account_update.ts
